requirements.txt
Pillow
torch
matplotlib
umap-learn
scikit-learn
torchvision
numpy
hdbscan
kneed
kagglehub
ipython
streamlit
pandas